+++ src/Ignixa.Application/Features/Mcp/Tools/PackageManagement/InstallPackageTool.cs
@@ -6,6 +6,7 @@
 using System.ComponentModel;
 using Microsoft.AspNetCore.Http;
 using ModelContextProtocol.Server;
+using Ignixa.Application.Features.Mcp.Dtos;
 using Ignixa.Domain.Abstractions;
 using Ignixa.PackageManagement.Abstractions;
 
@@ -112,65 +113,4 @@ NOTE: This operation may take 30-60 seconds for large packages.")]
                       $"({result.ImportedResources} new, {result.UpdatedResources} updated)"
         };
     }
-}
-
-/// <summary>
-/// Result of installing a package.
-/// </summary>
-public record InstallPackageResultDto
-{
-    /// <summary>
-    /// Whether the installation was successful.
-    /// </summary>
-    public required bool Success { get; init; }
-
-    /// <summary>
-    /// Tenant ID where package was installed.
-    /// </summary>
-    public required int TenantId { get; init; }
-
-    /// <summary>
-    /// Tenant name.
-    /// </summary>
-    public required string TenantName { get; init; }
-
-    /// <summary>
-    /// Package ID that was installed.
-    /// </summary>
-    public required string PackageId { get; init; }
-
-    /// <summary>
-    /// Package version that was installed.
-    /// </summary>
-    public required string PackageVersion { get; init; }
-
-    /// <summary>
-    /// Total number of resources extracted from package.
-    /// </summary>
-    public required int TotalResources { get; init; }
-
-    /// <summary>
-    /// Number of new resources imported.
-    /// </summary>
-    public required int ImportedResources { get; init; }
-
-    /// <summary>
-    /// Number of existing resources updated.
-    /// </summary>
-    public required int UpdatedResources { get; init; }
-
-    /// <summary>
-    /// Duration of installation in seconds.
-    /// </summary>
-    public required int DurationSeconds { get; init; }
-
-    /// <summary>
-    /// Breakdown of resources by type.
-    /// </summary>
-    public required Dictionary<string, int> ResourcesByType { get; init; }
-
-    /// <summary>
-    /// Human-readable message summarizing the installation.
-    /// </summary>
-    public required string Message { get; init; }
 }
